Placed video related tests into their own file
